+++ tests/conftest.py
@@ -9,6 +9,6 @@ import pytest
 def seed_random():
     init1 = random.SystemRandom().randrange(0, 2**32)
     init2 = random.SystemRandom().randrange(0, 2**32)
-    print(f"Ranom seeds: random.seed({init1}); np.random.seed({init2})")
+    print(f"Random seeds: random.seed({init1}); np.random.seed({init2})")
     random.seed(init1)
     np.random.seed(init2)